+++ api/src/opentrons/protocol_api/geometry.py
@@ -28,6 +28,7 @@ class CalibrationPosition:
     position: List[float]
     displayName: str
 
+
 class LabwareHeightError(Exception):
     pass
